requirements.txt
matplotlib==3.10.3
numpy==2.2.6
pandas==2.3.0
pillow==11.2.1
tqdm==4.67.1
albumentations==2.0.8
scikit-learn==1.7.0
timm==1.0.15
torch==2.7.1
transformers==4.52.4